medbot_backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Database
psycopg2-binary==2.9.7
django-extensions==3.2.3

# API Documentation
drf-spectacular==0.26.5

# Environment Management
python-decouple==3.8
django-environ==0.11.2

# Caching and Background Tasks
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Storage
boto3==1.29.7
django-storages==1.14.2
Pillow==10.1.0

# HTTP Requests for n8n integration
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Validation and Serialization
marshmallow==3.20.1
django-filter==23.3

# Security
cryptography==41.0.7
django-ratelimit==4.1.0

# Monitoring and Logging
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Development
ipython==8.17.2
django-silk==5.0.4
